Fix issues in API endpoints causing FATAL errors
Several missing `$id` parameters in controller functions caused
fatal undefined errors to be returned in the API response from
Slim.

The way user accounts were being serialized in a collection was
causing FATAL errors because each `User` was being converted to
an array and then an attempt was made to access the
`learningSkills` and `teachingSkills` properties of the object.

Fix uses of `users` table which is actually named `user`.