Fix composer.lock and re-seed command
The recent oauth commit contains some new libraries in composer.json,
but it did not include an updated composer.lock.  I have provided this
file in this commit.

In addition to the new libraries, that commit also re-named one of the
environment settings from development to dev.  This breaks the seed-db
phing task.  I have updated the task to use the appropriate environment.